typechecker: restructure and split up some functions
